PMC: switch python inspection to type() over isinstance()
The isinstance() matches all subclasses, but type() is a direct class match.
We dont want to screw up registered conversions for classes with wierd inheritance.
Instead this assumes the direct approach for default types.
If you want a subclass to be recognized, then manually add it,
or write an inspection routine that uses isinstance.
Its just that the default registrations wont.